PIPELINE: Refactor tests
