+++ HoloPipelines/core/services/np_image_manipulation.py
@@ -11,7 +11,10 @@ from config import INPUT_RESOLUTION_MAX
 
 import warnings
 
-warnings.filterwarnings("ignore", ".*the output shape of zoom() is calculated with*")
+# Ignore scipy/ndimage/interpolation.py:611: UserWarning: From scipy 0.13.0, the output
+# shape of zoom() is calculated with round() instead of int() - for thes e inputs the
+# size of the returned array has changed.
+warnings.filterwarnings("ignore", ".*the output shape of zoom.*")
 
 # FIXME: This is broken and needs to be fixed @Udomkarn
